requirements.txt
langchain
langchain-groq
duckduckgo-search
beautifulsoup4
requests
python-dotenv